requirements.txt
# watchdog
# apscheduler
# pydantic
opencv-python>=4.7.0.71
fastapi 
uvicorn>=0.20.0
pillow>=9.4.0
python-multipart
PyYAML
dependency_injector>=4.41.0
# torch>=1.13.1
# torchvision>=0.14.1
autologging
numpy>=1.23.5
pandas>=1.5.3
